requirements.txt
# 第1章用パッケージ

# データ処理・分析
pandas>=1.5.0
numpy>=1.21.0

# 設定ファイル読み込み
PyYAML>=6.0

# データ可視化
matplotlib>=3.5.0
seaborn>=0.11.0

# Jupyter Notebook (学習用)
jupyter>=1.0.0
ipykernel>=6.0.0

# 日付処理
python-dateutil>=2.8.0